back-end/requirements.txt
alembic==1.0.2
aniso8601==7.0.0
arrow==0.13.1
asn1crypto==0.24.0
Babel==2.6.0
blinker==1.4
certifi==2018.10.15
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.0
coverage==4.5.2
cryptography==2.7
elasticsearch==7.0.0
entrypoints==0.3
flake8==3.7.8
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Cors==3.0.6
Flask-HTTPAuth==3.2.4
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
httpie==0.9.9
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2019.1
redis==3.2.1
requests==2.20.0
rq==1.0
rq-dashboard==0.4.0
six==1.11.0
SQLAlchemy==1.2.13
urllib3==1.24
Werkzeug==0.14.1
